Error log backend is now its own class
